requirements.txt
flask==2.0.1
selenium==4.1.0
Pillow==9.0.0
pytesseract==0.3.8
openai==0.27.0
pandas==1.4.0
easyocr==1.4.2
python-dotenv==0.19.0
gunicorn==20.1.0
werkzeug==2.0.1 